# src/resources/swagger.yaml
openapi: 3.0.0
info:
  title: PPP Local Business API
  description: API that helps connect local business and their customers.
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  # Autenticação
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Registro de usuários
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                role:
                  type: string
              required:
                - username
                - password
                - role
      responses:
        '201':
          description: Usuário registrado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  username:
                    type: string
                  role:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
        '400':
          description: Dados inválidos
        '409':
          description: Usuário já existe
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login de usuários
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: Login bem-sucedido
        '401':
          description: Credenciais inválidas

  # Usuários
  /users:
    get:
      tags:
        - Users
      summary: Lista todos os usuários registrados
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de todos os usuários
        '403':
          description: Acesso negado

  # Categorias
  /business/categories:
    post:
      tags:
        - Business
      summary: Registro de categorias
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required:
                - name
      responses:
        '201':
          description: Categoria criada com sucesso
        '400':
          description: Dados inválidos
        '403':
          description: Acesso negado
        '409':
          description: Categoria já existe
    put:
      tags:
        - Business
      summary: Edição de categorias
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
              required:
                - id
                - name
      responses:
        '200':
          description: Categoria atualizada com sucesso
        '400':
          description: Dados inválidos
        '403':
          description: Acesso negado
        '404':
          description: Categoria não encontrada
    delete:
      tags:
        - Business
      summary: Remoção de categorias
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Categoria removida com sucesso
        '403':
          description: Acesso negado
        '404':
          description: Categoria não encontrada 
    get:
      tags:
        - Business
      summary: Lista todas as categorias
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de todas as categorias
        '403':
          description: Acesso negado
  # Empreendimentos
  /business/businesses:
    post:
      tags:
        - Business
      summary: Registro de empreendimentos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                category:
                  type: string
                description:
                  type: string
                isStarred:
                  type: boolean
                contact:
                  type: string
                  description: Informações de contato do empreendimento.
              required:
                - name
                - category
                - description
                - contact
      responses:
        '201':
          description: Empreendimento registrado com sucesso
        '400':
          description: Dados inválidos
        '403':
          description: Acesso negado
    put:
      tags:
        - Business
      summary: Edição de empreendimentos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
                category:
                  type: string
                description:
                  type: string
                isStarred:
                  type: boolean
                contact:
                  type: string
                  description: Informações de contato do empreendimento.
              required:
                - id
                - name
                - category
                - contact
      responses:
        '200':
          description: Empreendimento atualizado com sucesso
        '400':
          description: Dados inválidos
        '403':
          description: Acesso negado
        '404':
          description: Empreendimento não encontrado
    delete:
      tags:
        - Business
      summary: Remoção de empreendimentos
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Empreendimento removido com sucesso
        '403':
          description: Acesso negado
        '404':
          description: Empreendimento não encontrado
  /business:
    get:
      tags:
        - Business
      summary: Lista todos os empreendimentos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de todos os empreendimentos
        '403':
          description: Acesso negado
  /business/by-category:
    get:
      tags:
        - Business
      summary: Lista empreendimentos por categoria
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Lista de empreendimentos por categoria
        '403':
          description: Acesso negado
        '404':
          description: Categoria não encontrada
  /business/by-name:
    get:
      tags:
        - Business
      summary: Busca empreendimentos por nome
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Lista de empreendimentos por nome
        '403':
          description: Acesso negado
        '404':
          description: Empreendimento não encontrado
  /business/starred:
    get:
      tags:
        - Business
      summary: Lista empreendimentos marcados como estrela
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de empreendimentos estrela
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    category:
                      type: string
                    description:
                      type: string
                    isStarred:
                      type: boolean
        '403':
          description: Acesso negado
  # Produtos
  /products:
    post:
      tags:
        - Products
      summary: Registro de produtos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                businessId:
                  type: integer
                  description: ID of the existing business. The business must be registered before adding products.
                name:
                  type: string
                price:
                  type: number
                isStarred:
                  type: boolean
              required:
                - businessId
                - name
                - price
      responses:
        '201':
          description: Produto registrado com sucesso
        '400':
          description: Dados inválidos
        '403':
          description: Acesso negado
    put:
      tags:
        - Products
      summary: Edição de produtos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                businessId:
                  type: integer
                name:
                  type: string
                price:
                  type: number
                isStarred:
                  type: boolean
              required:
                - id
                - businessId
                - name
                - price
      responses:
        '200':
          description: Produto atualizado com sucesso
        '400':
          description: Dados inválidos
        '403':
          description: Acesso negado
        '404':
          description: Produto não encontrado
    delete:
      tags:
        - Products
      summary: Remoção de produtos
      security:
        - bearerAuth: []
      parameters:
        - name: businessId
          in: query
          schema:
            type: integer
          required: true
          description: ID do empreendimento ao qual o produto pertence.
        - name: id
          in: query
          schema:
            type: integer
          required: true
          description: ID do produto a ser removido.
      responses:
        '200':
          description: Produto removido com sucesso
        '403':
          description: Acesso negado
        '404':
          description: Produto não encontrado
    get:
      tags:
        - Products
      summary: Lista todos os produtos cadastrados
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de todos os produtos
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    businessId:
                      type: integer
                      description: ID do empreendimento ao qual o produto pertence.
                    name:
                      type: string
                    price:
                      type: number
                    isStarred:
                      type: boolean
        '403':
          description: Acesso negado
  /products/starred:
    get:
      tags:
        - Products
      summary: Busca produtos estrela (favoritos dos clientes)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de produtos estrela
        '403':
          description: Acesso negado
  /products/by-name:
    get:
      tags:
        - Products
      summary: Busca de produto por nome
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Lista de produtos por nome
        '403':
          description: Acesso negado
        '404':
          description: Produto não encontrado
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT